+++ src/lib.rs
@@ -14,9 +14,6 @@
 #![cfg_attr(feature="dev", feature(plugin))]
 #![cfg_attr(feature="dev", plugin(clippy))]
 
-#![feature(test)]
-extern crate test;
-
 #[macro_use]
 extern crate quick_error;
 extern crate protobuf;
